+++ src/types/observablearray.ts
@@ -7,6 +7,8 @@ import {isSpyEnabled, spyReportStart, spyReportEnd} from "../core/spy";
 import {arrayAsIterator, declareIterator} from "../utils/iterable";
 import {IEnhancer} from "../types/modifiers";
 
+const MAX_SPLICE_SIZE = 10000; // See e.g. https://github.com/mobxjs/mobx/issues/859
+
 // Detects bug in safari 9.1.1 (or iOS 9 safari mobile). See #364
 const safariPrototypeSetterInheritanceBug = (() => {
 	let v = false;
@@ -118,8 +120,12 @@ class ObservableArrayAdministration<T> implements IInterceptable<IArrayWillChang
 		let currentLength = this.values.length;
 		if (newLength === currentLength)
 			return;
-		else if (newLength > currentLength)
-			this.spliceWithArray(currentLength, 0, new Array(newLength - currentLength));
+		else if (newLength > currentLength) {
+			const newItems = new Array(newLength - currentLength);
+			for (let i = 0; i < newLength - currentLength; i++)
+				newItems[i] = undefined; // No Array.fill everywhere...
+			this.spliceWithArray(currentLength, 0, newItems);
+		}
 		else
 			this.spliceWithArray(newLength, currentLength - newLength);
 	}
@@ -171,11 +177,21 @@ class ObservableArrayAdministration<T> implements IInterceptable<IArrayWillChang
 		newItems = <T[]> newItems.map(v => this.enhancer(v, undefined));
 		const lengthDelta = newItems.length - deleteCount;
 		this.updateArrayLength(length, lengthDelta); // create or remove new entries
-		const res: T[] = this.values.splice(index, deleteCount, ...newItems); // FIXME: splat might exceed callstack size!
+		const res = this.spliceItemsIntoValues(index, deleteCount, newItems);
 
 		if (deleteCount !== 0 || newItems.length !== 0)
 			this.notifyArraySplice(index, newItems, res);
 		return res;
+	}
+
+	spliceItemsIntoValues(index, deleteCount, newItems: T[]): T[] {
+		if (newItems.length < MAX_SPLICE_SIZE) {
+			return this.values.splice(index, deleteCount, ...newItems);
+		} else {
+			const res = this.values.slice(index, index + deleteCount)
+			this.values = this.values.slice(0, index).concat(newItems, this.values.slice(index + deleteCount ))
+			return res;
+		}
 	}
 
 	notifyArrayChildUpdate<T>(index: number, newValue: T, oldValue: T) {